+++ src/arch/x86_64/interrupts/idt.rs
@@ -7,15 +7,19 @@ to appropriate sub-handlers that can be dynamically changed by the kernel.
 
 The common_irq_handler will get the irq number, and will have a dynamic dispatch table
 to call the appropriate handler for that irq. Each irq number could contain up to four handlers,
-as defined in the kernel/interrupts/definitions.rs file.
+as defined in the kernel/interrupts/consts file.
 */
-use lazy_static::lazy_static;
-use x86_64::structures::idt::{InterruptDescriptorTable, InterruptStackFrame};
-use crate::interrupts::dispatch_table::{IrqDispatchTable};
-use spin::{Once, Mutex};
-use paste::paste;
 use crate::arch::x86_64::interrupts::initializations::PICS;
 use crate::arch::x86_64::interrupts::initializations::PIC_1_OFFSET;
+use crate::interrupts::consts::HANDLERS_PER_IRQ_NUMBER;
+use crate::interrupts::irq_dispatch_table::IrqDispatchTable;
+use lazy_static::lazy_static;
+use paste::paste;
+use spin::{Mutex, Once};
+use x86_64::structures::idt::{InterruptDescriptorTable, InterruptStackFrame};
+
+/// Total number of IRQs supported.
+pub(super) const TOTAL_IRQS_NUMBER: usize = 16;
 
 lazy_static! {
     /// The Interrupt Descriptor Table (IDT) for the system.
@@ -24,7 +28,7 @@ lazy_static! {
 
     /// The IRQ Dispatch Table that holds the registered IRQ handlers.
     /// This table allows dynamic registration and unregistration of IRQ handlers.
-    pub static ref DISPATCH_TABLE: Mutex<IrqDispatchTable> = Mutex::new(IrqDispatchTable::new());
+    pub static ref DISPATCH_TABLE: Mutex<IrqDispatchTable<HANDLERS_PER_IRQ_NUMBER, TOTAL_IRQS_NUMBER>> = Mutex::new(IrqDispatchTable::new());
 }
 
 /// Generate automatically IRQ handlers for IRQs 0 to 15
@@ -56,8 +60,8 @@ generate_irq_handlers!(32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 4
 /// * `stack_frame` - The interrupt stack frame containing CPU state at the time of the interrupt.
 /// * `irq_num` - The IRQ number of the interrupt being handled.
 pub(super) fn common_irq_handler(stack_frame: InterruptStackFrame, irq_num: usize) {
-    DISPATCH_TABLE.lock().entries[irq_num - PIC_1_OFFSET as usize].run_handler(&stack_frame);
-    unsafe{PICS.lock().notify_end_of_interrupt(irq_num as u8)};
+    DISPATCH_TABLE.lock().run_handler(irq_num - PIC_1_OFFSET as usize, &stack_frame);
+    unsafe { PICS.lock().notify_end_of_interrupt(irq_num as u8) };
 }
 
 /// Initialize the Interrupt Descriptor Table (IDT) with IRQ handlers.
@@ -66,7 +70,7 @@ pub(super) fn common_irq_handler(stack_frame: InterruptStackFrame, irq_num: usiz
 pub(super) fn init_idt() {
     let mut idt = InterruptDescriptorTable::new();
 
-    for i in 0..IRQ_HANDLERS.len(){
+    for i in 0..IRQ_HANDLERS.len() {
         let vector_index = i as u8 + PIC_1_OFFSET;
         idt[vector_index].set_handler_fn(IRQ_HANDLERS[i]);
     }
