+++ src/interrupts/mod.rs
@@ -1,9 +1,10 @@
 use crate::arch::interrupts::DISPATCH_TABLE;
-use crate::interrupts::definitions::{IrqHandlerFn, IrqHandlerName};
 
-pub mod definitions;
-pub mod dispatch_table;
-
+pub mod consts;
+pub mod irq_dispatch_table;
+pub mod irq_handler_types;
+use irq_handler_types::{IrqHandlerFn, IrqHandlerName};
+mod irq_handlers_table_entry;
 /* the files here (definitions, dispatch_table) are somewhat of a mess. The files needs to be
 organized better, and I need to define what is in each file.
  */
